Move msm-related functionality out of `DensePolynomial` and into msm mod
