+++ jolt-core/src/poly/dense_mlpoly.rs
@@ -1,4 +1,5 @@
 #![allow(clippy::too_many_arguments)]
+use crate::msm::{flags_msm, sm_msm};
 use crate::poly::eq_poly::EqPolynomial;
 use crate::utils::{self, compute_dotproduct, compute_dotproduct_low_optimized, mul_0_1_optimized};
 
@@ -219,61 +220,15 @@ impl<F: PrimeField> DensePolynomial<F> {
                 let scalars = self.Z[R_size * i..R_size * (i + 1)].as_ref();
                 match hint {
                     CommitHint::Normal => Commitments::batch_commit(scalars, &gens),
-                    CommitHint::Flags => Self::flags_msm(scalars, &gens),
+                    CommitHint::Flags => flags_msm(scalars, &gens),
                     CommitHint::Small => {
                         let bigints: Vec<_> = scalars.iter().map(|s| s.into_bigint()).collect();
-                        Self::sm_msm(&bigints, &gens)
+                        sm_msm(&bigints, &gens)
                     }
                 }
             })
             .collect();
         PolyCommitment { C }
-    }
-
-    /// Special MSM where all scalar values are 0 / 1 – does not verify.
-    fn flags_msm<G: CurveGroup>(scalars: &[G::ScalarField], bases: &[G::Affine]) -> G {
-        assert_eq!(scalars.len(), bases.len());
-        let result = scalars
-            .into_iter()
-            .enumerate()
-            .filter(|(_index, scalar)| !scalar.is_zero())
-            .map(|(index, scalar)| bases[index])
-            .sum();
-
-        result
-    }
-
-    pub fn sm_msm<V: VariableBaseMSM>(
-        scalars: &[<V::ScalarField as PrimeField>::BigInt],
-        bases: &[V::MulBase],
-    ) -> V {
-        assert_eq!(scalars.len(), bases.len());
-        let num_buckets: usize = 1 << 16; // TODO(sragss): This should be passed in / dependent on M = N^{1/C}
-
-        // #[cfg(test)]
-        // scalars.for_each(|scalar| {
-        //     assert!(scalar < V::ScalarField::from(num_buckets as u64).into_bigint())
-        // });
-
-        // Assign things to buckets based on the scalar
-        let mut buckets: Vec<V> = vec![V::zero(); num_buckets];
-        scalars.into_iter().enumerate().for_each(|(index, scalar)| {
-            let bucket_index: u64 = scalar.as_ref()[0];
-            buckets[bucket_index as usize] += bases[index];
-        });
-
-        let mut result = V::zero();
-        let mut running_sum = V::zero();
-        buckets
-            .into_iter()
-            .skip(1)
-            .enumerate()
-            .rev()
-            .for_each(|(index, bucket)| {
-                running_sum += bucket;
-                result += running_sum;
-            });
-        result
     }
 
     #[tracing::instrument(skip_all, name = "DensePolynomial.bound")]
@@ -995,32 +950,6 @@ mod tests {
         );
     }
 
-    #[test]
-    fn sm_msm_parity() {
-        use ark_curve25519::{EdwardsAffine as G1Affine, EdwardsProjective as G1Projective, Fr};
-        let mut rng = test_rng();
-        let bases = vec![
-            G1Affine::rand(&mut rng),
-            G1Affine::rand(&mut rng),
-            G1Affine::rand(&mut rng),
-        ];
-        let scalars = vec![Fr::from(3), Fr::from(2), Fr::from(1)];
-        let expected_result = bases[0] + bases[0] + bases[0] + bases[1] + bases[1] + bases[2];
-        assert_eq!(bases[0] + bases[0] + bases[0], bases[0] * scalars[0]);
-        let expected_result_b =
-            bases[0] * scalars[0] + bases[1] * scalars[1] + bases[2] * scalars[2];
-        assert_eq!(expected_result, expected_result_b);
-
-        let calc_result_a: G1Projective = VariableBaseMSM::msm(&bases, &scalars).unwrap();
-        assert_eq!(calc_result_a, expected_result);
-
-        let scalars_bigint: Vec<_> = scalars
-            .into_iter()
-            .map(|scalar| scalar.into_bigint())
-            .collect();
-        let calc_result_b: G1Projective = DensePolynomial::<Fr>::sm_msm(&scalars_bigint, &bases);
-        assert_eq!(calc_result_b, expected_result);
-    }
 
     #[test]
     fn commit_with_hint_parity() {
